Agentic_Ai-main/Agentic_Ai-main/Day 7/Multi Agent (LangGraph)/main/requirements.txt
langgraph
langchain
openai
python-dotenv
tqdm
requests
beautifulsoup4
serpapi
streamlit
PyMuPDF
langchain
sentence-transformers
faiss-cpu
